+++ packages/terra-table/src/Table.jsx
@@ -101,6 +101,11 @@ const propTypes = {
    * rendered to allow for row selection to occur.
    */
   hasSelectableRows: PropTypes.bool,
+
+  /**
+   * Boolean specifying whether or not the table should have zebra striping for rows.
+   */
+  isStriped: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -131,6 +136,7 @@ function Table(props) {
     onColumnSelect,
     onCellSelect,
     hasSelectableRows,
+    isStriped,
     rowHeaderIndex,
   } = props;
 
@@ -312,6 +318,7 @@ function Table(props) {
                 rowHeaderIndex={rowHeaderIndex}
                 onCellSelect={isGridContext ? handleCellSelection : undefined}
                 isSelected={row.isSelected}
+                isTableStriped={isStriped}
               />
             ))}
           </tbody>