+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 
 * Added
+  * Added test to cover enabling zebra striping for the `terra-table` component.
+
+* Updated
+  * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.
+
+## 1.41.0 - (October 17, 2023)
+
+* Added
   * Added documentation for the new `terra-table` component.
   
 * Changed
@@ -12,11 +20,15 @@
   * Updates examples for `terra-data-grid` to cover scenarios for focusable elements prevented from receiving focus.
   * Updated focusable cell test for `terra-data-grid` to account for focusable elements in a hidden container.
 
+* Added
+ * Added example for vertical orientation functionality for `terra-tabs`.
+
 ## 1.40.0 - (October 11, 2023)
 
 * Added
   * Added documentation for the new `terra-file-path` component.
   * Added documentation for FlowsheetDataGrid in `terra-data-grid`.
+  * Added additional documentation and implementation guide in adherence to the newly established documentation template for the examples of Worklist Data Grid in `terra-data-grid`.
 
 ## 1.39.0 - (October 3, 2023)